requirements.txt
Flask==2.0.2
Flask-WTF==0.15.1
Flask-Login==0.5.1
dash==1.21.0
dash-bootstrap-components==0.15.0
mysql-connector-python==8.0.26
dash-core-components==2.0.0
dash-html-components==2.0.0
sentence-transformers==2.1.0
pandas==1.3.3
